refactor: renaming config_classes to config
