+++ cognite/powerops/resync/config/cogshop/shop_file_config.py
