+++ cognite/powerops/resync/config/production/connections.py
