+++ cognite/powerops/resync/config/shared.py
