+++ cognite/powerops/resync/to_models/to_cogshop_model.py
@@ -16,7 +16,7 @@ from cognite.powerops.clients.cogshop.data_classes import (
     ModelTemplateApply,
     TransformationApply,
 )
-from cognite.powerops.resync.config_classes.resync_config import CogShopConfig
+from cognite.powerops.resync.config.resync_config import CogShopConfig
 from cognite.powerops.resync.models import cogshop
 from cognite.powerops.resync.models.cdf_resources import CDFSequence
 from cognite.powerops.resync.models.cogshop import CogShopModel
