+++ cognite/powerops/resync/to_models/to_production_model.py
@@ -6,8 +6,8 @@ from typing import Optional
 import pandas as pd
 from cognite.client.data_classes import Sequence, TimeSeries
 
-from cognite.powerops.resync.config_classes.production.connections import Connection
-from cognite.powerops.resync.config_classes.resync_config import ProductionConfig
+from cognite.powerops.resync.config.production.connections import Connection
+from cognite.powerops.resync.config.resync_config import ProductionConfig
 from cognite.powerops.resync.models import production
 from cognite.powerops.resync.models.cdf_resources import CDFSequence
 from cognite.powerops.resync.utils.serializer import load_yaml
